context: extract the std trait implementations into a separate module
Mostly because this way I can ignore these in test coverage.
