+++ Makefile
@@ -37,6 +37,7 @@ RS_OBJECTS = \
 	src/cache_yamls.rs \
 	src/cache_yamls/tests.rs \
 	src/context.rs \
+	src/context/system.rs \
 	src/context/tests.rs \
 	src/cron.rs \
 	src/cron/tests.rs \
@@ -108,6 +109,9 @@ build: $(RS_OBJECTS) Cargo.toml Makefile
 # Without coverage: cargo test --lib
 check-unit: Cargo.toml $(RS_OBJECTS) locale/hu/LC_MESSAGES/osm-gimmisn.mo data/yamls.cache
 	cargo tarpaulin --lib -v --skip-clean --fail-under 100 --target-dir ${PWD}/target-cov ${CARGO_OPTIONS}
+
+check-unit-llvm: Cargo.toml $(RS_OBJECTS) locale/hu/LC_MESSAGES/osm-gimmisn.mo data/yamls.cache
+	cargo llvm-cov --lib -q --ignore-filename-regex system.rs --fail-under-lines 100
 
 src/browser/config.ts: wsgi.ini Makefile
 	printf 'const uriPrefix = "%s";\nexport { uriPrefix };\n' $(shell grep prefix wsgi.ini |sed 's/uri_prefix = //') > $@
